rearrange report test reports for more validation
We're going to start restricting what can be submitted through this API
to make our job easier on the backend: Clients must submit clean data,
so we don't have to clean it up ourselves.

To do this, we need to start making examples of invalid data, so we need
a place to put them.