do not allow "v" prefix in Perl 5 versions
Metabase clients currently send a report from Perl 5.26.0 as "v5.26.0".
The "v" is not a crucial part of the version, so it's removed by the
backend to normalize everything.

As Metabase clients get moved to the new format, we need to make sure
they submit valid data. Preventing them from adding "v" at the beginning
of the version number will simplify our work on the backend.