fix error after successful metabase registration
If a user tried registering with the Metabase, they would get an error
message response. However, their user would be successfully created
anyway, and they'd be able to submit reports as normal.

It seems a Metabase registration has multiple parts, and I was only
handling one of the parts correctly. I now check to see if the fact is
like I expect it, and if it isn't, I warn and log what it is so that
I can see what's happening and deal with it later.